az-python-automation/requirements.txt
azure-identity
azure-keyvault-secrets
azure-mgmt-resource